Update main.cpp
Attempts to compare strings and display a percentage of how accurate the string is. Hopefully it meets the criteria of being case sensitive and displays the right percentage but it has been a while since I've done string comparison and I'm rusty on how to do it.  I tried my best...